+++ tests/test-add-admin-css.php
@@ -20,6 +20,7 @@ class Add_Admin_CSS_Test extends WP_UnitTestCase {
 
 		if ( class_exists( 'c2c_AddAdminCSS' ) ) {
 			c2c_AddAdminCSS::instance()->reset();
+			unset( $_GET[ c2c_AddAdminCSS::NO_CSS_QUERY_PARAM ] );
 		}
 	}
 
@@ -284,6 +285,34 @@ class Add_Admin_CSS_Test extends WP_UnitTestCase {
 
 		return $out;
 	}
+
+	public function test_remove_query_param_from_redirects() {
+		$url = 'https://example.com/wp-admin/themes.php?page=add-admin-css%2Fadd-admin.css.php';
+
+		$this->assertEquals(
+			$url,
+			c2c_AddAdminCSS::instance()->remove_query_param_from_redirects( $url . '&' . c2c_AddAdminCSS::NO_CSS_QUERY_PARAM . '=1' )
+		);
+	}
+
+	public function test_can_show_css() {
+		$this->assertTrue( c2c_AddAdminCSS::instance()->can_show_css() );
+	}
+
+	public function test_can_show_css_with_false_query_param() {
+		$_GET[ c2c_AddAdminCSS::NO_CSS_QUERY_PARAM ] = '1';
+
+		$this->assertFalse( c2c_AddAdminCSS::instance()->can_show_css() );
+	}
+
+	public function test_recovery_mode_via_query_param_disables_add_css() {
+		$this->test_can_show_css_with_false_query_param();
+
+		$out = $this->test_add_css_to_head_with_just_css( '' );
+
+		$this->assertEmpty( $out );
+	}
+
 
 	/*
 	 * Setting handling